Split PayChan into PayChanCreate, PayChanFund, PayChanClaim
